+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 
 
+## [1.1.3] - 2025-03-07
+
+#### Added
+- Adds function to generate json file for AlphaFold3 input using FlashFold MSA.
+
+#### Changed
+- Version number is updated.
+- Python scripts have been relocated to the `tools` directory. 
+
+#### Fixed
+- NA
+
+#### Removed
+- NA
+
+
 ## [1.1.2] - 2025-03-04
 
 #### Added